This is using the Manhattan distance instead of the euclidean distance. Perfomance is worse.
