+++ Utilities.py
@@ -1,0 +1,50 @@
+def path_printer(maze, path):
+    print(" ", end=" ")
+    for row in range(len(maze)): print(row, end = " ")
+    print("")
+    for i in range(len(maze)):  # Iterate over rows
+        print(i, end = " ")
+        for j in range(len(maze[i])):  # Iterate over columns
+            if (i, j) in path:
+                print("X", end=" ")
+            else:
+                print("O", end=" ")
+        print()  # Print a newline after each row
+
+def maze_chooser (i):
+    if i == 1:
+        maze = [[0, 0, 0, 0, 1, 0, 0, 0, 0, 0],
+                [0, 0, 0, 0, 1, 0, 0, 0, 0, 0],
+                [0, 0, 0, 0, 1, 0, 0, 0, 0, 0],
+                [0, 0, 0, 0, 1, 0, 0, 0, 0, 0],
+                [0, 0, 0, 0, 1, 0, 0, 0, 0, 0],
+                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+                [0, 0, 0, 0, 1, 0, 0, 0, 0, 0],
+                [0, 0, 0, 0, 1, 0, 0, 0, 0, 0],
+                [0, 0, 0, 0, 1, 0, 0, 0, 0, 0],
+                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0]]
+        return maze
+    elif i == 2:
+        maze = [[1, 1, 1, 1, 0, 1, 1, 1, 1, 1],
+                [1, 1, 1, 1, 0, 1, 1, 1, 1, 1],
+                [1, 1, 1, 1, 0, 1, 1, 1, 1, 1],
+                [1, 1, 1, 1, 0, 1, 1, 1, 1, 1],
+                [1, 1, 1, 1, 0, 1, 1, 1, 1, 1],
+                [1, 1, 1, 1, 0, 1, 1, 1, 1, 1],
+                [1, 1, 1, 1, 0, 1, 1, 1, 1, 1],
+                [1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
+                [1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
+                [1, 1, 1, 1, 1, 1, 1, 1, 1, 1]]
+        return maze
+    else:
+        maze = [[1, 1, 1, 0, 1, 1, 1, 1, 1, 1],
+                [1, 1, 1, 0, 1, 1, 1, 1, 1, 1],
+                [1, 1, 1, 0, 1, 1, 1, 1, 1, 1],
+                [1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
+                [1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
+                [1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
+                [1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
+                [1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
+                [1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
+                [1, 1, 1, 1, 1, 1, 1, 1, 1, 1]]
+        return maze
